Move GameController
